feat(front): create separate features for tauri and browser backends. move tauri backend code into s separate module;
