+++ totes/Cargo.toml
@@ -5,14 +5,19 @@ authors = ["Pavlo Myroniuk <the.best.tvarynka@gmail.com>"]
 edition = "2021"
 repository = "https://github.com/TheBestTvarynka/totes"
 
+[features]
+default = ["tauri_backend"]
+tauri_backend = ["dep:wasm-bindgen", "dep:wasm-bindgen-futures", "dep:serde-wasm-bindgen"]
+browser_backend = []
+
 [dependencies]
 leptos = { version = "0.6", features = ["csr"] }
 leptos_hotkeys = "0.2"
-wasm-bindgen = { version = "0.2" }
-wasm-bindgen-futures = "0.4"
+wasm-bindgen = { version = "0.2", optional = true }
+wasm-bindgen-futures = { version = "0.4", optional = true }
 js-sys = "0.3"
 serde = { version = "1", features = ["derive"] }
-serde-wasm-bindgen = "0.6"
+serde-wasm-bindgen = { version = "0.6", optional = true }
 console_error_panic_hook = "0.1"
 log = "0.4"
 wasm-logger = "0.2"
